+++ src/lib/connection/XMPPServiceHandler/alertEventHandler.ts
@@ -385,6 +385,9 @@ class AlertEventHandler extends GenericHandler {
                     case "poll":
                         // treated in conversationEventHandler
                         break;
+                    case "joincompanyinvite":
+                        // treated in invitationEventHandler
+                        break;
                     default:
                         that.logger.log("error", LOG_ID + "(onManagementMessageReceived) unmanaged management message node " + node.getName());
                         break;