+++ src/lib/connection/XMPPServiceHandler/channelEventHandler.ts
@@ -157,6 +157,9 @@ class ChannelEventHandler extends GenericHandler {
                     case "poll":
                         // treated in conversationEventHandler
                         break;
+                    case "joincompanyinvite":
+                        // treated in invitationEventHandler
+                        break;
                     default:
                         that.logger.log("error", LOG_ID + "(onManagementMessageReceived) unmanaged management message node " + node.getName());
                         break;